requirements.txt
RPi.bme280
smbus2
cryptography==2.5
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth==1.6.2
google-cloud-pubsub==0.39.1
pyjwt==1.7.1
requests==2.21.0
